PAI optimize (#3067)
* pai optimize

* fix import

* fix ut
